pyproject.toml
[tool.poetry]
name = "flowselect"
version = "0.1.0"
description = "Code for running experiments from 'Normalizing Flows for Knockoff-free Controlled Feature Selection'."
authors = ["Anonymous"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
torch = "^1.7.1"
hydra-core = "^1.0.6"
pandas = "^1.2.1"
numpy = "~1.19"
pytorch-lightning = "^1.1.7"
pylint = "^2.6.0"
sklearn = "^0.0"
seaborn = "^0.11.1"
h5py = "~2.10"
black = "^20.8b1"
pytest = "^6.2.2"
tensorboard = "^2.4.1"
pytest-cov = "^2.11.1"
rpy2 = "^3.4.2"
nflows = "^0.14"
tensorflow = "^2.4.1"
cvxpy = "^1.1.11"
cvxopt = "^1.2.6"
einops = "^0.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
